CustomMessageBox: add overrides

+++ Utilities/AltitudeAngel/AltitudeAngel.cs
@@ -31,7 +31,7 @@ namespace MissionPlanner.Utilities.AltitudeAngel
             }
             if (CustomMessageBox.Show(
                     "Do you wish to enable Altitude Angel airspace management data?\nFor more information visit [link;http://www.altitudeangel.com;www.altitudeangel.com]",
-                    "Altitude Angel - Enable", CustomMessageBox.MessageBoxButtons.YesNo) == (int)CustomMessageBox.DialogResult.Yes)
+                    "Altitude Angel - Enable", CustomMessageBox.MessageBoxButtons.YesNo) == CustomMessageBox.DialogResult.Yes)
             {
                 await service.SignInAsync();
             }
